Refactor galaxy.tool_shed.tool_shed_registry -> galaxy.util.tool_shed.
common_util makes heavy use of this implicitly so typing and designwise this will be better.
